Refactor to remove ConnectedPolygon.js
